Fix compatibility with Groovy 3.0.9
